+++ Sources/TuistKit/Services/ScaffoldService.swift
@@ -2,6 +2,7 @@ import TSCBasic
 import TuistCore
 import TuistGraph
 import TuistLoader
+import TuistPlugin
 import TuistScaffold
 import TuistSupport
 
@@ -13,14 +14,17 @@ enum ScaffoldServiceError: FatalError, Equatable {
         }
     }
 
-    case templateNotFound(String)
+    case templateNotFound(String, searchPaths: [AbsolutePath])
     case nonEmptyDirectory(AbsolutePath)
     case attributeNotProvided(String)
 
     var description: String {
         switch self {
-        case let .templateNotFound(template):
-            return "Could not find template \(template). Make sure it exists at Tuist/Templates/\(template)"
+        case let .templateNotFound(template, searchPaths):
+            let searchDirectories = searchPaths
+                .reduce(into: Set<AbsolutePath>()) { result, path in result.insert(path.parentDirectory) }
+                .reduce("\n") { $0 + "  * \($1.pathString)\n" }
+            return "Could not find template \(template) in: \(searchDirectories)"
         case let .nonEmptyDirectory(path):
             return "Can't generate a template in the non-empty directory at path \(path.pathString)."
         case let .attributeNotProvided(name):
@@ -29,30 +33,39 @@ enum ScaffoldServiceError: FatalError, Equatable {
     }
 }
 
-class ScaffoldService {
+final class ScaffoldService {
     private let templateLoader: TemplateLoading
     private let templatesDirectoryLocator: TemplatesDirectoryLocating
     private let templateGenerator: TemplateGenerating
+    private let configLoader: ConfigLoading
+    private let pluginService: PluginServicing
 
-    init(templateLoader: TemplateLoading = TemplateLoader(),
-         templatesDirectoryLocator: TemplatesDirectoryLocating = TemplatesDirectoryLocator(),
-         templateGenerator: TemplateGenerating = TemplateGenerator())
-    {
+    init(
+        templateLoader: TemplateLoading = TemplateLoader(),
+        templatesDirectoryLocator: TemplatesDirectoryLocating = TemplatesDirectoryLocator(),
+        templateGenerator: TemplateGenerating = TemplateGenerator(),
+        configLoader: ConfigLoading = ConfigLoader(manifestLoader: ManifestLoader()),
+        pluginService: PluginServicing = PluginService()
+    ) {
         self.templateLoader = templateLoader
         self.templatesDirectoryLocator = templatesDirectoryLocator
         self.templateGenerator = templateGenerator
+        self.configLoader = configLoader
+        self.pluginService = pluginService
     }
 
-    func loadTemplateOptions(templateName: String,
-                             path: String?) throws -> (
+    func loadTemplateOptions(
+        templateName: String,
+        path: String?
+    ) throws -> (
         required: [String],
         optional: [String]
     ) {
         let path = self.path(path)
-        let directories = try templatesDirectoryLocator.templateDirectories(at: path)
-
+        let plugins = try loadPlugins(at: path)
+        let templateDirectories = try locateTemplateDirectories(at: path, plugins: plugins)
         let templateDirectory = try self.templateDirectory(
-            templateDirectories: directories,
+            templateDirectories: templateDirectories,
             template: templateName
         )
 
@@ -68,14 +81,15 @@ class ScaffoldService {
         }
     }
 
-    func run(path: String?,
-             templateName: String,
-             requiredTemplateOptions: [String: String],
-             optionalTemplateOptions: [String: String?]) throws
-    {
+    func run(
+        path: String?,
+        templateName: String,
+        requiredTemplateOptions: [String: String],
+        optionalTemplateOptions: [String: String?]
+    ) throws {
         let path = self.path(path)
-
-        let templateDirectories = try templatesDirectoryLocator.templateDirectories(at: path)
+        let plugins = try loadPlugins(at: path)
+        let templateDirectories = try locateTemplateDirectories(at: path, plugins: plugins)
 
         let templateDirectory = try self.templateDirectory(
             templateDirectories: templateDirectories,
@@ -109,6 +123,11 @@ class ScaffoldService {
         }
     }
 
+    private func loadPlugins(at path: AbsolutePath) throws -> Plugins {
+        let config = try configLoader.loadConfig(path: path)
+        return try pluginService.loadPlugins(using: config)
+    }
+
     /// Parses all `attributes` from `template`
     /// If those attributes are optional, they default to `default` if not provided
     /// - Returns: Array of parsed attributes
@@ -136,6 +155,18 @@ class ScaffoldService {
         }
     }
 
+    /// Locates all template directories, local, system, and plugin.
+    /// - Parameter path: The path where the command is being executed.
+    /// - Returns: A list of template directories
+    private func locateTemplateDirectories(
+        at path: AbsolutePath,
+        plugins: Plugins
+    ) throws -> [AbsolutePath] {
+        let templateRelativeDirectories = try templatesDirectoryLocator.templateDirectories(at: path)
+        let templatePluginDirectories = plugins.templateDirectories
+        return templateRelativeDirectories + templatePluginDirectories
+    }
+
     /// Finds template directory
     /// - Parameters:
     ///     - templateDirectories: Paths of available templates
@@ -144,7 +175,7 @@ class ScaffoldService {
     private func templateDirectory(templateDirectories: [AbsolutePath], template: String) throws -> AbsolutePath {
         guard
             let templateDirectory = templateDirectories.first(where: { $0.basename == template })
-        else { throw ScaffoldServiceError.templateNotFound(template) }
+        else { throw ScaffoldServiceError.templateNotFound(template, searchPaths: templateDirectories) }
         return templateDirectory
     }
 }