+++ Tests/Fixtures/WorkspaceWithPlugins/FrameworkA/Project.swift
@@ -1,0 +1,14 @@
+import LocalPlugin
+import ProjectDescription
+
+let project = Project(
+    name: "FrameworkA",
+    targets: [
+        .framework(
+            name: "FrameworkA",
+            dependencies: [
+                .project(target: "FrameworkB", path: "//FrameworkB"),
+            ]
+        ),
+    ]
+)