+++ Tests/Fixtures/WorkspaceWithPlugins/FrameworkB/Project.swift
@@ -1,0 +1,11 @@
+import LocalPlugin
+import ProjectDescription
+
+let project = Project(
+    name: "FrameworkB",
+    targets: [
+        .framework(
+            name: "FrameworkB"
+        ),
+    ]
+)