+++ projects/website/src/components/layout.jsx
@@ -2,11 +2,9 @@
 
 import { jsx } from 'theme-ui'
 import React from 'react'
-import { Styled } from 'theme-ui'
 import GlobalStyle from './global-style'
 import Header from './header'
 import Footer from './footer'
-import CookieBanner from './cookie-banner'
 import { Helmet } from 'react-helmet'
 
 const Layout = ({ children, menuOpen, setMenuOpen, menuRef }) => {
@@ -23,7 +21,6 @@ const Layout = ({ children, menuOpen, setMenuOpen, menuRef }) => {
       <GlobalStyle />
       <Header menuOpen={menuOpen} setMenuOpen={setMenuOpen} menuRef={menuRef} />
       <main sx={{ mb: [3, 6] }}>{children}</main>
-      <CookieBanner />
       <Footer />
     </>
   )